requirements.txt
numpy==1.19.5
torch==1.8.0+cu101
torchsummary==1.5.1
Pillow==8.3.1
